:coffin:

+++ src/views/AccountView.vue
@@ -3,24 +3,19 @@
     <div class="row">
       <h1 class="account-heading animate__animated animate__fadeInDown">Account</h1>
     </div>
-    <div class="row justify-content-center" v-if="user">
-      <Card class="card">
-        <template #cardHeader>
+    
+    <div class="row" v-if="user">
+      <div class="account-details-container">
+        <div class="img-container">
           <img :src="user.userProfile" loading="lazy" class="img-fluid img-width" :alt="user.firstName">
-        </template>
-        <template #cardBody>
-          <h3 class="card-title fw-bold">{{ user.firstName }}</h3>
-          <h3 class="card-title fw-bold">{{ user.lastName }}</h3>
+        </div>
+        <div class="account-detail">
+          <h3 class="fw-bold">First Name: {{ user.firstName }}</h3>
+          <h3 class="fw-bold">Last Name: {{ user.lastName }}</h3>
           <p class="lead"><span class="fw-bold">Email Address</span>: {{ user.emailAdd }}</p>
           <p class="lead"><span class="fw-bold">Age</span>: {{ user.userAge }}</p>
           <p class="lead"><span class="fw-bold">Gender</span>: {{ user.Gender }}</p>
-        </template>
-      </Card> 
-    </div>
-    <div v-else>
-      <Spinner />
-    </div>
-    <div class="accountButtons">
+          <div class="accountButtons">
         <button class="account-button" data-bs-toggle="modal" :data-bs-target="'#updateAccountModal' + user.userID"><i class="bi bi-pen-fill"></i></button>
         <button class="account-button" @click="deleteUser(user.userID)"><i class="bi bi-trash-fill"></i></button>
       </div>
@@ -28,26 +23,31 @@
         <router-link to="/"><button class="btnHome">Go back home</button></router-link>
         <router-link to="/logout">
 
-          <button class="btnHome">Logout</button>
+          <button class="btnHome" :disabled="isUserDeleted">Logout</button>
         </router-link>
       </div>
-      
+        </div>
+      </div>
+    </div>
+    <div v-else>
+      <Spinner />
+    </div>
+       
     <UpdateAccountModal :user="user" @update="handleUpdate" />
   </div>
 </template>
 
 <script setup>
 import { useStore } from 'vuex'
-import { computed, onMounted, watch } from 'vue'
-import Card from '@/components/Card.vue'
+import { ref, computed, onMounted, watch } from 'vue'
 import Spinner from '@/components/Spinner.vue'
 import UpdateAccountModal from '@/components/UpdateAccountModal.vue'
-// import { useRoute } from 'vue-router'
 import { useCookies } from 'vue3-cookies'
+import Swal from 'sweetalert2'
 
 const store = useStore()
-// const route = useRoute()
 const { cookies } = useCookies()
+const isUserDeleted = ref(false) // Variable to track if the user is deleted
 
 const user = computed(() => store.state.user || cookies.get('LegitUser') )
 
@@ -67,8 +67,28 @@ function handleUpdate() {
 }
 
 function deleteUser(userID) {
-store.dispatch('deleteUser', userID);
-cookies.remove('LegitUser')
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#FF9A00',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      store.dispatch('deleteUser', userID)
+      cookies.remove('LegitUser')
+      isUserDeleted.value = true // Disable logout button when user is deleted
+
+      Swal.fire({
+      title: 'Deleted!',
+      text: 'Your account has been deleted.',
+      icon: 'success',
+      confirmButtonColor: '#FF9A00'
+    })
+    }
+  })
 }
 
 watch(user, (newUser) => {
@@ -98,17 +118,11 @@ background-color: var(--awesome);
   padding-bottom: 2rem;
 }
 
-.card{
-  font-family: "Poppins", sans-serif;
-  color: var(--primary);
-  width: 25%;
-}
-
 .accountButtons{
   display: flex;
   justify-content: space-between;
   margin: auto;
-  width: 24%;
+  width: 100%;
 }
 
 .btnHome{
@@ -131,17 +145,47 @@ background-color: var(--awesome);
 display: flex;
 justify-content: center;
 flex-direction: column;
-width: 24%;
+width: 100%;
 margin: auto;
 gap: 1rem;
-margin-top: 2rem;
+margin-top: 8rem;
+}
+
+.account-details-container{
+  display: flex;
+  justify-content: space-evenly;
+  width: 90%;
+  margin: auto;
+  font-family: "Poppins", sans-serif;
+  color: var(--primary);
+  padding-top: 2rem;
+}
+
+.img-container{
+  width: 30%;
+} 
+
+.img-container img{
+  border-radius: 0.3rem;
+  box-shadow: 0.8rem 0.5rem 1rem var(--borderColor);
 }
 
 @media (width < 999px)
 {
-.card{
-  width: 85%;
+  .account-details-container{
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  width: 100%;
+  margin: auto;
+  padding-top: 0rem;
 }
+
+.img-container{
+  width: 85%;
+  padding-bottom: 3rem;
+  margin: auto;
+} 
 
 .accountButtons{
   display: flex;
